Split tests to match code organization
Reorganized test files to follow the same structure as the implementation:

- main_test.go: Tests for main.go (run(), usage(), subcommand dispatch)
- validate_test.go: Tests for validate.go (validateInvoice(), formatters, runValidate())

Added new test coverage:
- TestRun: Tests command-line argument handling
- TestUsage: Verifies main usage output
- TestSubcommandDispatch: Tests subcommand routing
- TestRunValidate: Tests validate command argument handling
- TestIntegration_ValidateCommand: End-to-end validation tests

All tests pass with improved organization and maintainability.
